Move central keml service into keml core
