+++ src/app/shared/keml/graphical/helper/information/info-choice/info-choice.component.ts
@@ -1,7 +1,7 @@
 import {Component, EventEmitter, Input, Output, TemplateRef, ViewChild} from '@angular/core';
 import {Information, Preknowledge, ReceiveMessage} from "@app/shared/keml/core/msg-info";
 import {MatDialog, MatDialogRef} from "@angular/material/dialog";
-import {KemlService} from "@app/features/editor/services/keml.service";
+import {KemlService} from "@app/shared/keml/core/keml.service";
 import { PreknowledgeComponent } from '../../../preknowledge/preknowledge.component';
 import { MsgComponent } from '../../../msg/msg.component';
 import { InfoInnerComponent } from '../info-inner/info-inner.component';
